backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
openai==1.57.2
anthropic==0.40.0
qdrant-client==1.12.1
numpy==1.26.4
pdfplumber==0.11.4
python-docx==1.1.2
